refactor(src): move pages to /src
