+++ package.json
@@ -4,10 +4,10 @@
   "private": false,
   "license": "Apache License 2.0",
   "scripts": {
-    "start": "next start",
-    "build": "next build",
-    "dev": "next dev",
-    "now-build": "next build",
+    "dev": "next dev ./src",
+    "start": "next start ./src",
+    "build": "next build ./src",
+    "now-build": "next build ./src",
     "deploy": "now",
     "release": "standard-version",
     "release:dry": "yarn release --dry-run",
