+++ src/pages/_document.js
@@ -1,7 +1,8 @@
 import Document, { Head, Main, NextScript } from 'next/document';
 import { ServerStyleSheets } from '@material-ui/styles';
 import flush from 'styled-jsx/server';
-import theme from '../src/theme';
+
+import theme from '../theme';
 
 class CustomDocument extends Document {
   render() {
