+++ src/static/apple-touch-icon-180x180.png
