+++ src/static/apple-touch-icon-60x60.png
