+++ src/static/apple-touch-icon-76x76.png
