+++ src/static/browserconfig.xml
